backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vidshield-ai-backend"
version = "0.1.0"
description = "VidShield AI — Backend API"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",

    # Database
    "sqlalchemy==2.0.35",
    "alembic==1.13.3",
    "asyncpg==0.29.0",          # async PostgreSQL driver
    "psycopg2-binary==2.9.9",   # sync driver for Alembic CLI

    # Pydantic & settings
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "pydantic[email]",

    # Auth
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.2.1",              # pin: passlib incompatible with bcrypt>=5.x

    # Cache / broker
    "redis[hiredis]==5.1.1",
    "celery[redis]==5.4.0",

    # Logging
    "structlog==24.4.0",

    # HTTP client (for webhooks, presigned URL verification)
    "httpx==0.27.2",

    # File / multipart
    "python-multipart==0.0.12",

    # AWS
    "boto3==1.35.28",

    # AI / ML
    "openai==1.51.0",
    "langchain==0.2.16",
    "langgraph==0.2.28",
    "langchain-openai==0.1.23",
    "pinecone>=3.0.0",

    # Video processing
    "opencv-python-headless==4.10.0.84",
    "yt-dlp>=2024.1.1",               # download videos from YouTube and 1000+ sites

    # WebSockets / Socket.IO
    "python-socketio[asyncio]==5.11.4",

    # Utilities
    "python-dateutil==2.9.0",
    "orjson==3.10.7",

    # PDF generation
    "reportlab==4.2.5",

    # Notifications
    "sendgrid>=6.11.0",
    "twilio>=9.0.0",

    # Payments
    "stripe>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "factory-boy==3.3.1",
    "aiosqlite==0.20.0",
    "ruff==0.6.9",
    "mypy==1.11.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
